Separate LargeListItemV from ListItemP
